requirements.txt
pillow

# wayland
dbus-python
pygobject

# X11
xlib
